+++ docs/README.md
@@ -9,6 +9,8 @@
   - 입력된 당첨 번호 및 보너스 번호를 관리하는 클래스
 [x] 로또 당첨 확인 클래스 
   - 승리 번호와 자동 생성된 티켓의 번호를 통해 당첨 결과를 확인하는 클래스
+[x] 순위 확인 클래스
+  - 상수로 구성하고, 순위를 확인하는 함수가 존재하는 클래스 
 
 ### 뷰
 - inputView
@@ -54,4 +56,5 @@
     - 45 이내의 숫자가 아닌 경우
     - 중복된 숫자가 포함된 경우
     - 숫자가 아닌 경우
+
