+++ src/main/java/lotto/domain/LottoMachine.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public class LottoMachine {
+}
