+++ src/main/java/lotto/domain/LottoResult.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public class LottoResult {
+}
