+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public class LottoTicket {
+}
