+++ src/main/java/lotto/domain/WinningLotto.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public class WinningLotto {
+}
