+++ src/main/java/lotto/exception/InvalidInputException.java
@@ -1,0 +1,4 @@
+package lotto.exception;
+
+public class InvalidInputException {
+}
