+++ src/main/java/lotto/service/LottoService.java
@@ -1,0 +1,4 @@
+package lotto.service;
+
+public class LottoService {
+}
