+++ src/main/java/lotto/utils/RandomsUtil.java
@@ -1,0 +1,4 @@
+package lotto.utils;
+
+public class RandomsUtil {
+}
