+++ src/main/java/lotto/view/InputView.java
@@ -1,0 +1,4 @@
+package lotto.view;
+
+public class InputView {
+}
